♻ Extract prettyBox to simpleTable component
